fix(cachemire): resolve switched calls in their own currency, never against the old expectation
Thomas caught the live line 'cache held · read 28.2k of 20.6k expected ·
prefix stayed warm' after a sol→fable switch: 20.6k is sol-billed, 28.2k is
fable-billed, and composing them claims an impossible 137% read. The same
flaw sat in classification (hit ratio across currencies), the partial render,
and restored-session ledgers.

- classifyCall: when the model switched, the call's own billed prompt is the
  denominator; the ratio becomes 'how much of this prompt was already cached'.
- CallRecord.switched marks records whose expectedRead is old-currency;
  renderHeldLine speaks target currency only ('cache held · read 28.2k
  (100% of prompt) · the new model already had the prefix cached') and the
  partial render says 'of X prompt' instead of 'of Y expected'.
- ledger restore detects identity changes between persisted calls (in the
  data, not invented), classifies in the new currency, and keeps 'model
  switched …' as the cause instead of 'restored (cause unknown)'.
- moved fingerprinting/diff/entry-match/classification into classify.ts
  (index.ts 1202 → 947 baseline ratchet); split ledger-restore tests out.
- smoke: tolerate both honest outcomes of the abort (switch miss, or warm
  hold via the aborted send's write) and compare forecast accuracy against
  the re-baselined panel total, never a stray old-currency re-wrote line.
  Both branches live-validated 16/16.
- design language §7 + pi-cachemire.md document the resolution currency rule.
