+++ scripts/dev/agent-lint-baseline.json
@@ -5,7 +5,7 @@
   "lineBudgets": {
     "defaultTsMax": 350,
     "files": {
-      "extensions/pi-cachemire/index.ts": 1202,
+      "extensions/pi-cachemire/index.ts": 947,
       "extensions/pi-contextimate/index.ts": 1549,
       "extensions/pi-traceline/index.ts": 1766,
       "tests/contract/pi-internals.test.ts": 432,
